refactor(server): split Discord adapter into 4 layers
